+++ crates/fuzzyfold/src/bin/ff-timecourse.rs
@@ -1,12 +1,16 @@
 use std::fs;
+use std::fs::File;
+use std::io::Write;
+use std::io::BufWriter;
 use std::sync::Arc;
 use std::path::Path;
 use std::path::PathBuf;
+
 use rayon::prelude::*;
 use rand::rng;
+use colored::*;
 use clap::Parser;
 use anyhow::Result;
-use colored::*;
 use indicatif::ProgressBar;
 use indicatif::ProgressStyle;
 use serde_json::to_string_pretty;
@@ -43,6 +47,9 @@ pub struct Cli {
     #[arg(short, long, value_name = "FILE")]
     output: PathBuf,
 
+    #[arg(short, long)]
+    title: Option<String>,
+
     #[command(flatten, next_help_heading = "Simulation parameters")]
     simulation: TimelineParameters,
 
@@ -58,7 +65,7 @@ fn main() -> Result<()> {
     cli.simulation.validate()?;
 
     // --- Build simulator ---
-    let emodel = cli.energy.build_model();
+    let emodel = Arc::new(cli.energy.build_model());
     let rmodel = cli.kinetics.build_model(emodel.temperature());
 
     let is_rna = cli.energy.dna.is_none();
@@ -74,17 +81,29 @@ fn main() -> Result<()> {
         cli.num_sims, cli.kinetics, cli.simulation, cli.energy);
 
     let times = cli.simulation.get_output_times();
-    let mut macrostates = MacrostateRegistry::from((&sequence, &emodel));
+    let mut macrostates = MacrostateRegistry::from((sequence.clone(), emodel.clone()));
     macrostates.insert_files(&cli.macrostates)?;
-
-    println!("Macrostates:\n{}", macrostates.iter()
-        .map(|(_, m)| format!(" - {} {:6.2}", m.name(), m.ensemble_energy().unwrap_or(0.0)))
-        .collect::<Vec<_>>().join("\n"));
-
+    // Verbose Output
+    println!("{:>4} {:<10} {} {:>5} {:>8}",
+        "ID",
+        "Macrostate".cyan(), format!("{}", sequence).yellow(), "Size", "Energy");
+    for (id, m) in macrostates.iter() {
+        if m.name() == "Unassigned" {
+            println!("{:4} {:<10}", 0, m.name());
+            continue
+        }
+        println!("{:4} {:<10} {:<} {:>5} {:>8.2}",
+            id, 
+            m.name(),
+            m.get_lowest_microstate().unwrap(),
+            m.len(),
+            m.ensemble_energy().unwrap());
+    }
     let shared_macrostates = Arc::new(macrostates);
 
     let tln_path = cli.output.with_extension("tln");
     let svg_path = cli.output.with_extension("svg");
+    let nxy_path = cli.output.with_extension("nxy");
 
     // If timeline.json exists, reload instead of starting empty
     let mut master = 
@@ -95,30 +114,29 @@ fn main() -> Result<()> {
             println!("A new timeline file will be created: {}", tln_path.display());
             Timeline::new(&times, Arc::clone(&shared_macrostates))
         };
-    let amodel = Arc::new(cli.energy.build_model());
 
     let timelines: Vec<_> =
         match (rmodel.k3ws().is_some(), rmodel.k4ws().is_some()) {
             (false, false) => {
-                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, amodel, NoShift))
+                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, emodel, NoShift))
                     .map_err(|e| anyhow::anyhow!("failed to construct AddDelMoves: {:?}", e))?;
                 run_timecourse(moves, rmodel, cli.simulation.t_end, cli.num_sims as u64,
                     Arc::clone(&shared_macrostates), &times).collect()
             },
             (true, false) => {
-                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, amodel, ThreeWayOnly))
+                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, emodel, ThreeWayOnly))
                     .map_err(|e| anyhow::anyhow!("failed to construct AddDelMoves: {:?}", e))?;
                 run_timecourse(moves, rmodel, cli.simulation.t_end, cli.num_sims as u64,
                     Arc::clone(&shared_macrostates), &times).collect()
             },
             (false, true) => {
-                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, amodel, FourWayOnly))
+                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, emodel, FourWayOnly))
                     .map_err(|e| anyhow::anyhow!("failed to construct AddDelMoves: {:?}", e))?;
                 run_timecourse(moves, rmodel, cli.simulation.t_end, cli.num_sims as u64,
                     Arc::clone(&shared_macrostates), &times).collect()
             },
             (true, true) => {
-                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, amodel, ThreeAndFour))
+                let moves = LoopNeighbors::try_from((sequence.clone(), &pairings, emodel, ThreeAndFour))
                     .map_err(|e| anyhow::anyhow!("failed to construct AddDelMoves: {:?}", e))?;
                 run_timecourse(moves, rmodel, cli.simulation.t_end, cli.num_sims as u64,
                     Arc::clone(&shared_macrostates), &times).collect()
@@ -129,31 +147,59 @@ fn main() -> Result<()> {
         master.merge(timeline);
     }
 
-    println!("Final Timeline:\n{}", master);
-    plot_occupancy_over_time(&master, svg_path, cli.simulation.t_ext, cli.simulation.t_end);
+    println!("{}", "Finished simulations!".red());
+
+    // save / print / plot.
+    let mut writer = BufWriter::new(File::create(nxy_path.clone())?);
+    write!(writer, "{}", master)?;
+    println!("Wrote nxy file: {}", format!("{}",nxy_path.display()).green());
+
     let serial = master.to_serializable();
     let json = to_string_pretty(&serial).unwrap();
-    fs::write(tln_path, json).unwrap();
+    fs::write(tln_path.clone(), json).unwrap();
+    println!("Wrote tln file: {}", tln_path.display());
+
+    let numsim = master.points[0].counter;
+    let title = cli.title.unwrap_or({
+        format!("ff-timecourse ({} simulations)", 
+        {
+            if numsim >= 10000 {
+                let s = numsim.to_string();
+                let mut out = String::new();
+                for (i, c) in s.chars().rev().enumerate() {
+                    if i > 0 && i % 3 == 0 {
+                        out.push('_');
+                    }
+                    out.push(c);
+                }
+                out.chars().rev().collect::<String>()
+            } else { 
+                numsim.to_string()
+            }
+        })
+    });
+
+    plot_occupancy_over_time(&master, svg_path.clone(), &title, cli.simulation.t_ext, cli.simulation.t_end);
+    println!("Plotted svg file: {}", svg_path.display());
 
     Ok(())
 }
 
 
-fn run_timecourse<'t, W, K, E>(
+fn run_timecourse<W, K, E>(
     moves: W,
     rmodel: K,
     t_end: f64,
     num_sims: u64,
-    registry: Arc<MacrostateRegistry<'t, E>>,
-    times: &'t [f64],
-) -> impl ParallelIterator<Item = Timeline<'t, E>>
+    registry: Arc<MacrostateRegistry<E>>,
+    times: &[f64],
+) -> impl ParallelIterator<Item = Timeline<E>>
 where
     W: Walker + Clone + Send + Sync,
-    K: RateModel + Sync + Clone + std::marker::Send,
-    E: EnergyModel + Sync,
+    K: RateModel + Clone,
+    E: EnergyModel,
     SSA<W, K>: From<(W, K)>,
 {
-    println!("Simulation progress:");
     let pb = ProgressBar::new(num_sims);
     pb.set_style(
         ProgressStyle::default_bar()