+++ crates/ff_structure/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this crate will be documented in this file.
 
+## [0.3.1] - 2026-01-13
+### Added
+- NAIDX indexing for PairTable
+
 ## [0.3.0] - 2026-01-13
 ### Added/Changed
 - multi_pair_table.rs: new structures for MultiPairTable & StrandPairTable